vision-box/backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 视觉检测核心库
supervision==0.16.0
ultralytics==8.0.206
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# 文件处理
python-multipart==0.0.6
aiofiles==23.2.1

# 工具库
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 日志和监控
loguru==0.7.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 打包工具
pyinstaller==6.2.0